norm: push_swap.h

+++ 05_Push_Swap/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: ottalhao <ottalhao@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/28 16:25:46 by ottalhao          #+#    #+#             */
-/*   Updated: 2026/01/04 11:58:43 by ottalhao         ###   ########.fr       */
+/*   Updated: 2026/01/04 14:44:09 by ottalhao         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,40 +16,40 @@
 # include <unistd.h>
 # include <stdlib.h>
 
-typedef struct	s_list
+typedef struct s_list
 {
 	int				n;
 	int				index;
 	struct s_list	*next;
-} t_list;
+}	t_list;
 
-void	ft_putstr_fd(char *s, int fd);
-long	ft_atol(const char *nptr);
-int		is_empty(char **asc_n);
-char	**ft_split(const char *s, char c);
-int		count_lst(t_list **lst);
-void	ft_lstadd_back(t_list **lst, t_list *node);
-t_list	*ft_lstnew(int n);
+void		ft_putstr_fd(char *s, int fd);
+long long	ft_atol(const char *nptr);
+int			is_empty(char **asc_n);
+char		**ft_split(const char *s, char c);
+int			count_lst(t_list **lst);
+void		ft_lstadd_back(t_list **lst, t_list *node);
+t_list		*ft_lstnew(int n);
 
-int		str_empty(char *s);
-void	rotate_stack(t_list **lst, char* operation);
-void	rotate_stack_both(t_list **stack_a, t_list **stack_b);
-void	rev_rotate_stack(t_list **lst, char *operation);
-void	rrr(t_list **stack_a, t_list **stack_b);
-void	swap_stack(t_list **lst, char *operation);
-void	swap_stack_both(t_list **stack_a, t_list **stack_b);
-void	push_stack(t_list **stack_a, t_list **stack_b, char *operation);
+int			str_empty(char *s);
+void		rotate_stack(t_list **lst, char *operation);
+void		rotate_stack_both(t_list **stack_a, t_list **stack_b);
+void		rev_rotate_stack(t_list **lst, char *operation);
+void		rrr(t_list **stack_a, t_list **stack_b);
+void		swap_stack(t_list **lst, char *operation);
+void		swap_stack_both(t_list **stack_a, t_list **stack_b);
+void		push_stack(t_list **stack_a, t_list **stack_b, char *operation);
 
-void	assign_indexes(t_list **lst);
-int		find_distance(t_list *lst, int n);
+void		assign_indexes(t_list **lst);
+int			find_distance(t_list *lst, int n);
 
-void	free_tab(char **tab);
-void	free_stack(t_list **lst);
-int		error_exit(t_list **a, t_list **b, char **tab);
+void		free_tab(char **tab);
+void		free_stack(t_list **lst);
+int			error_exit(t_list **a, t_list **b, char **tab);
 
-void	pswp_sort_3(t_list **stack_a);
-void	pswp_sort_4(t_list **stack_a, t_list **stack_b);
-void	pswp_sort_5(t_list **stack_a, t_list **stack_b);
-void	pswp_sort(t_list **stack_a, t_list **stack_b, unsigned int count);
+void		pswp_sort_3(t_list **stack_a);
+void		pswp_sort_4(t_list **stack_a, t_list **stack_b);
+void		pswp_sort_5(t_list **stack_a, t_list **stack_b);
+void		pswp_sort(t_list **stack_a, t_list **stack_b, unsigned int count);
 
 #endif
